requirements.txt
Pillow>=10.0.0
PyYAML>=6.0
pydantic>=2.0,<3.0
requests>=2.31.0
urllib3<2
pytest>=7.4
